Выделил абстрактный класс и применил наследование.
