+++ src/task-component.js
@@ -1,0 +1,33 @@
+import {createElement} from './utils';
+
+export default class TaskComponent {
+  constructor() {
+    if (new.target === TaskComponent) {
+      throw new Error(`Can't instantiate Component, only concrete one.`);
+    }
+    this._element = null;
+    this._state = {};
+  }
+  get element() {
+    return this._element;
+  }
+
+  get template() {
+    throw new Error(`You have to define template.`);
+  }
+
+  render() {
+    this._element = createElement(this.template);
+    this.bind();
+    return this._element;
+  }
+
+  bind() {}
+
+  unbind() {}
+
+  unrender() {
+    this.unbind();
+    this._element = null;
+  }
+}
